Dashboard and verification

+++ app/Http/Controllers/API/JobApiController.php
@@ -10,79 +10,74 @@ use Illuminate\Http\Request;
 
 class JobApiController extends Controller
 {
-    /**
-     * Display a listing of the resource.
-     *
-     * @return \Illuminate\Http\Response
-     */
+
     public function index()
     {
         //retrieve all jobs
         try {
             $jobs = Job::all();
+            if (!$jobs) {
+                return Api_Utils::error("Jobs not found", 404);
+            }
             return Api_Utils::success($jobs, "Jobs successfully returned", 200);
         } catch (\Exception $e) {
             return Api_Utils::error($e->getMessage(), 400);
         }
     }
 
-    /**
-     * Store a newly created resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @return \Illuminate\Http\Response
-     */
     public function store(JobStoreRequest $request)
     {
         //create jobs
         try {
             $job = Job::create($request->all());
-            return Api_Utils::success($job, "Job successfully created", 201);
+            if ($job) {
+                return Api_Utils::success($job, "Job successfully created", 201);
+            }
+            return Api_Utils::error("Job not created", 400);
         } catch (\Exception $e) {
             return Api_Utils::error($e->getMessage(), 400);
         }
     }
 
-    /**
-     * Display the specified resource.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
     public function show($id)
     {
         //show a job
         try {
             $job = Job::find($id);
+            if (!$job) {
+                return Api_Utils::error("Job not found", 404);
+            }
             return Api_Utils::success($job, "Job successfully returned", 200);
         } catch (\Exception $e) {
             return Api_Utils::error($e->getMessage(), 400);
         }
     }
 
-    /**
-     * Update the specified resource in storage.
-     *
-     * @param  \Illuminate\Http\Request  $request
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function update(Request $request, $id)
     {
-        //
+        //edit a job
+        try {
+            $job = Job::find($id);
+            if (!$job) {
+                return Api_Utils::error("Job not found", 404);
+            }
+            $job->update($request->all());
+            return Api_Utils::success($job, "Job successfully updated", 200);
+        } catch (\Exception $e) {
+            return Api_Utils::error($e->getMessage(), 400);
+        }
     }
 
-    /**
-     * Remove the specified resource from storage.
-     *
-     * @param  int  $id
-     * @return \Illuminate\Http\Response
-     */
+
     public function destroy($id)
     {
         //delete a job
         try {
             $job = Job::find($id);
+            if (!$job) {
+                return Api_Utils::error("Job not found", 404);
+            }
             $job->delete();
             return Api_Utils::success($job, "Job successfully deleted", 200);
         } catch (\Exception $e) {
